@@ -15,45 +15,46 @@
       "test:unit": "vitest"
    },
    "devDependencies": {
-      "@sveltejs/adapter-auto": "^3.0.0",
-      "@sveltejs/kit": "^2.5.22",
-      "@sveltejs/vite-plugin-svelte": "^3.1.0",
+      "@sveltejs/adapter-auto": "^3.3.1",
+      "@sveltejs/kit": "^2.8.4",
+      "@sveltejs/vite-plugin-svelte": "^4.0.2",
       "@types/jsonwebtoken": "^9.0.6",
       "@typescript-eslint/eslint-plugin": "^6.0.0",
       "@typescript-eslint/parser": "^6.0.0",
-      "autoprefixer": "^10.4.16",
+      "autoprefixer": "^10.4.20",
       "eslint": "^9.3.0",
       "eslint-config-prettier": "^9.1.0",
       "eslint-plugin-svelte": "^2.36.0-next.13",
-      "postcss": "^8.4.32",
-      "prettier": "^3.1.1",
+      "prettier": "^3.3.3",
       "prettier-plugin-svelte": "^3.1.2",
-      "svelte": "^5.0.0-next.220",
-      "svelte-check": "^3.6.0",
-      "svelte-motion": "^0.12.1",
-      "tailwindcss": "^3.4.1",
-      "tslib": "^2.4.1",
-      "typescript": "^5.0.0",
-      "vite": "^5.0.0",
+      "svelte": "^5.2.8",
+      "svelte-check": "^4.1.0",
+      "svelte-motion": "^0.12.2",
+      "tailwindcss": "^3.4.15",
+      "tslib": "^2.8.1",
+      "typescript": "^5.7.2",
+      "vite": "^5.4.11",
       "vitest": "^1.0.0"
    },
    "type": "module",
    "dependencies": {
-      "@auth/sveltekit": "^1.4.1",
-      "@fontsource-variable/inter": "^5.0.16",
-      "@formkit/auto-animate": "^0.8.1",
-      "@microsoft/signalr": "^8.0.0",
-      "axios": "^1.6.2",
-      "clsx": "^2.1.0",
-      "dexie": "^3.2.4",
+      "@auth/sveltekit": "^1.7.4",
+      "@fontsource-variable/inter": "^5.1.0",
+      "@formkit/auto-animate": "^0.8.2",
+      "@microsoft/signalr": "^8.0.7",
+      "axios": "^1.7.8",
+      "bits-ui": "^0.21.16",
+      "clsx": "^2.1.1",
+      "dexie": "^4.0.10",
       "jsonwebtoken": "^9.0.2",
-      "lucide-svelte": "^0.311.0",
+      "lucide-svelte": "^0.461.0",
       "peerjs": "^1.5.4",
-      "svelte-sonner": "^0.3.11",
+      "postcss": "^8.4.49",
+      "svelte-sonner": "^0.3.28",
       "tailwind-merge": "^2.2.0",
       "tailwind-variants": "^0.1.20",
-      "ulidx": "^2.2.1",
+      "ulidx": "^2.4.1",
       "wavesurfer.js": "^7.8.2",
-      "zod": "^3.22.4"
+      "zod": "^3.23.8"
    }
 }
