@@ -1,13 +1,13 @@
 <script lang="ts">
-   import { Header } from '$lib/application/header';
    import NavigationBar from '$lib/application/navigation-bar/index.svelte';
+
+   let { children } = $props();
 </script>
 
 <div class="max-h-screen h-screen">
-   <Header />
-
-   <div class="flex w-full max-h-[calc(100%-3.75rem)] h-full">
+   <div class="flex w-full max-h-[100%] h-full">
       <NavigationBar />
-      <slot />
+
+      {@render children()}
    </div>
 </div>
